updated model for pin, seeder for areas

+++ db/models/pin.js
@@ -11,8 +11,8 @@ module.exports = (sequelize, DataTypes) => {
   }
   Pin.init(
     {
-      lat: DataTypes.INTEGER,
-      lng: DataTypes.INTEGER,
+      lat: DataTypes.FLOAT,
+      lng: DataTypes.FLOAT,
       placeName: DataTypes.STRING,
       areaId: {
         type: DataTypes.INTEGER,
